Updates from PR comments
- Moved activity walking functions from lara api file to the existing activity utils file
- Replaced temp activity variable with additional if check
