+++ src/lara-api.ts
@@ -1,6 +1,6 @@
 import { sampleActivities, sampleSequences } from "./data";
 import { Activity, Sequence } from "./types";
-import { runningInCypress } from "./utilities/cypress";
+import { rewriteModelsResourcesUrls } from "./utilities/activity-utils";
 
 export const getActivityDefinition = (activity: string): Promise<Activity> => {
   return new Promise((resolve, reject) => {
@@ -15,56 +15,6 @@ export const getActivityDefinition = (activity: string): Promise<Activity> => {
       }
     }
   });
-};
-
-const walkActivity = (activityNode: any, stringCallback: (s: string) => string) => {
-  if (!activityNode) {
-    return;
-  }
-  if (activityNode instanceof Array) {
-    for (const i in activityNode) {
-      walkActivity(activityNode[i], stringCallback);
-    }
-  } else if (typeof activityNode === "object" ) {
-    Object.keys(activityNode).forEach(key => {
-      switch (typeof activityNode[key]) {
-        case "string":
-          activityNode[key] = stringCallback(activityNode[key]);
-          break;
-        case "object":
-          walkActivity(activityNode[key], stringCallback);
-          break;
-      }
-    });
-  } else if (typeof activityNode === "string") {
-    activityNode = stringCallback(activityNode);
-  }
-};
-
-const rewriteModelsResourcesUrls = (activity: Activity) => {
-  // do not rewrite urls when running in Cypress, otherwise the sample activity iframes do not load causing timeouts
-  if (runningInCypress) {
-    return activity;
-  }
-
-  walkActivity(activity, (s) => {
-    return s
-      .replace(/https?:\/\/models-resources\.concord\.org/, "models-resources")
-      .replace(/https?:\/\/models-resources\.s3\.amazonaws\.com/, "models-resources")
-      .replace(/https?:\/\/((.+)-plugin)\.concord\.org/, "models-resources/$1");
-
-  });
-  return activity;
-};
-
-export const getAllUrlsInActivity = (activity: Activity, urls: string[] = []) => {
-  walkActivity(activity, (s) => {
-    if (/^(\s*https?:\/\/|models-resources)/.test(s)) {
-      urls.push(s);
-    }
-    return s;
-  });
-  return urls;
 };
 
 const getActivityDefinitionFromLara = (activityUrl: string): Promise<Activity> => {
